Transformed login into API
